requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Vector Database
chromadb==0.4.22

# Ollama
ollama==0.1.6

# Utils
python-dotenv==1.0.0
pydantic==2.5.3